+++ changelog.md
@@ -11,8 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- SHA-256 checksum verification for downloads to ensure integrity.
+- New `bvm stats` command to show statistics about the BoxLang installation, including the number of installed modules, their versions, and the total size of the installation.
 - More internal docs
 - install-bx-module helper tests
+- Network connection checker for bvm to assist on disconnected systems
+- Global cleanup helper in case something goes wrong during version installations
 
 ### Changed
 
@@ -20,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Invalid usage of `mktemp` by not using `xxx` placeholders, which caused issues on some systems.
 - Fixed the `install-bx-module` script to use the right helpers path.
 - Build script was not updating the rigth `@build.version` so it was never updated in the bvm and installer.'